Add typehints for template variables
The single biggest deficiency in SMR design is that templates are not
able to know what variables are being populated into their environment
by the Template class. As a result, static analysis assumes that all
variables are undefined (unless defined locally), and therefore also
has no knowledge about their types. Therefore, any changes to the code
require manual testing of potentially every page.

There are so many pages, and so many different possible states for each
page, that it is not reasonable to write tests for them all (and, in
fact, we have _zero_ template tests right now). Static analysis is the
only way that we can have any confidence in our changes.

The two most likely solutions to this issue are:

1. Promote template pages to proper classes/functions, so that their
   interface is well-defined, and so that inputs can be validated in
   both the caller and callee contexts.

2. Convert from PHP templating to a first-class templating library,
   such as Twig or Latte. Some PHPStan extensions exist for this job,
   but it's unclear how mature/well-maintained they are (see
   https://github.com/deprecated-packages/phpstan-twig-rules and
   https://github.com/efabrica-team/phpstan-latte for example).

Either option will be a massive refactor, and I'm hesitant to commit to
one or the other at that point. In the meantime, we need a stop-gap
measure, and that's what we implement here.

We add `@var` declarations for every variable in the templates so that
we at least can resolve the "undefined variable" PHPStan warnings. This
allows us to define the type of every expected variable as well and fix
internal inconsistencies and incorrect APIs.

What this does _not_ do is ensure that the Page classes populating the
Template data are passing in consistent objects, since the `@var`
declarations are taken by PHPStan as truth. Therefore, it is possible
for Page classes and templates to get out of sync.