Add TradeGoodTransaction helper class
For the purposes of path finding with Plotter, we need a container to
hold both a TradeGood and a TransactionType. We replace a generic array
with a TradeGoodTransaction class to more robustly pass around and
identify this container.

Since this significantly simplifies the SmrPort::hasX method, we can
remove it in favor of performing the check in SmrSector::hasX.