phpcs.xml: add UnusedVariable sniff
Adds sniff SlevomatCodingStandard.Variables.UnusedVariable.

In most cases, this is just removing unused array key definitions in
foreach loops. We allow array values to be defined even if they are
unused, because `$array as $key => $value` is the fastest way to loop
over an array (it's about 50% faster than using `array_keys`).