Request: add method getBool
This allows us to easily convert boolean-like request variables (e.g.
"true/false", 0/1, "yes/no", etc.) to PHP booleans.

For now, this will throw on non-boolean-like inputs, because none of
our forms should allow users to make this mistake. In the future, we
may want to perform similar validation to int/float (with filter_var),
but we'd need to first improve form validation for these types.