Remove Scrutinizer CI
We are at the point where PHPStan and PHP_CodeSniffer outperform
Scrutinizer CI in every way. The problems with Scrutinizer are as
follows:

* Closed source
* Runs outside the context of the rest of our CI (GHA)
* Analysis is sometimes significantly delayed
* No support for PHP 8+ language features (abandonware?)

The remaining warnings on Scrutinizer are all related to failures in
parsing PHP 8+. It's been a great (though sometimes frustrating) tool
these past few years, but it's time to move on.