PlottedCourse.inc.php: fix invalid SmrSector call
This was using the wrong interface to SmrSector::getSector, resulting
in forcing an update of the cache unexpectedly due to the implicit type
conversion from int to bool. This bug is as old as the repo.

Now that we are using strict types, this invalid call throws, and so we
fix it to prevent the erroneous forced update.

Fixes PHPStan warning:

> Parameter #3 $forceUpdate of static method SmrSector::getSector() expects bool, int given.