PlotCourse: move logic out of template into Page
To avoid intermixing logic and HTML, we move the aggregation of plot
options into the PlotCourse page class. We can use a simple array
keyed by ID with name values to use to construct the HTML. This allows
us to use `asort` on the array instead of using a complicated Sorter
class method.