Remove Sorter class in favor of uasort calls
The convenience of using builtin sorting methods has improved a lot in
PHP 7+ with arrow functions and spaceship operators. Therefore, we can
forego the complicated Sorter class in favor of direct `uasort` calls.

We use `uasort`, even when we don't necessarily need to maintain index
association (i.e. we could use `usort` instead), just to avoid future
misuse of the sorted array, since the keys of most of our arrays are
meaningful (account IDs, etc.).

The downside of using `uasort` over Sorter is that we need to specify
the sorting method twice (once for each object in the comparison).