phpstan.neon.dist: update template ignores
We are now defining most variables in `src/templates`, though we have
not yet tackled the `admin` templates, so we need to continue to ignore
undefined variables there.

A few other warnings arose as a result of properly defining all the
template variables, which cannot be resolved without refactoring or by
fixes to PHPStan upstream.