SectorMap.inc.php: more robust $Port definition
The previous logic for defining `$Port` was correct, but it was not
sufficiently robust for PHPStan to know incontrovertibly that it will
be defined when used.

By tightening up the logic, we can guarantee that `$Port` is defined
before it is used.

I think that declaring `Sector::hasCachedPort` pure may have had a
similar impact, but this new logic is better anyway because it is
less redundant.