Add TakenDamageData type aliases
Whereas `WeaponDamageData` holds the info about damage that _can_ be
done by a weapon, `{Force}TakenDamageData` is for damage that _was_
done to a combatant. We use a type alias for convenience, though this
should probably be a readonly data class in the future.

The `Force::takeDamageTo{Mines,CDs,SDs}` methods now return the number
of forces destroyed, rather than the total damage done. This is so that
we can avoid division when recording the `Num{Mines,CDs,SDs}` to ensure
that they are identified as ints instead of floats by PHPStan.