Mines: don't mutate TakenDamageData type
The number of mines launched is reduced based on the actual damage that
they deal to their target (to prevent, e.g., single shield clearing
having the same effectiveness as a full hit).

This was being tracked by adding a `Launched` field to TakenDamageData,
which mutates the array in a way that we can't properly handle.

Since we're actually changing the number that launch, it's just as
accurate (if not more accurate) to modify the `Launched` field of the
WeaponDamageData-type array.