Add array shapes to combat return types
The returned arrays in combat methods are monstrous and need to be
converted into type aliases or data classes. This will be challenging
because the combat code needs to be refactored so that there aren't
trivial differences between each of the returned arrays (e.g. using
needlessly specific field names like "TargetPort" and "TargetPlanet").

This will be an expensive undertaking. Before this can be done, it is
important to have very specific validation of the array structure so
that when we do refactor it, we can be sure it is done correctly.

With the explicit array shapes, we need to add some assertions to
placate PHPStan, which doesn't understand that, e.g., if "CheckHit" is
true, then "ActualDamage" and "WeaponDamage" must exist.

The `Player::shoot*` methods are removed, because they are trivial
wrappers around the `Ship::shoot*` methods, and the ship is the real
combat entity.