DiContainer: mutate ContainerBuilder when compiling
While the original `$builder` object is identical to the return value
of `enableCompilation`, the function can mutate the template type to
that of the `containerParentClass`. By replacing `$builder` with the
return value (a no-op, logically), it informs PHPStan that the template
type of `$builder` has changed.

Note that if PHP-DI used `@phpstan-self-out self<T>` to annotate the
`enableCompilation` function, this workaround would not be needed.

Fixes PHPStan warning:

> Method Smr\Container\DiContainer::buildContainer() never returns
> Smr\Container\ResettableCompiledContainer so it can be removed from
> the return type.