Rework handling of Session errors in ajax calls
Convert Session errors during ajax requests from UserError exceptions
to immediate exits.

The reason for this is that these caused pressing the back button to
immediately redirect to an error page. This presumably defeated the
purpose of using the back button (which is usually to look at pages
that you previously visited).

An alternative use of the back button is to navigate to another page,
and in this case it would be better for the pages to continue ajax
updating so all links are refreshed (and then maybe we wouldn't even
need ReusableTrait?!). But this seems less common, and is more prone
to error, especially if any of the previous pages are non-reusable.
This is the tradeoff for being able to properly view previous pages.

I tried using JS to disable ajax when a back button press was detected,
so that the error messages could be retained for cases outside the
expected "back button" behavior, but I could not find any simple and
robust method to do this.

Note that this is a reversal of 8710eebda4 based on my experience with
trying to navigate around since that commit.