Add more strict-rules checks
These are all the remaining "strict-rules" checks (though not all of
the base PHPStan stricter analysis checks).

Add OVERRIDE_FORWARD to the dynamicConstantNames list. Even though it
is only ever set to true if it is set at all, we do want to ensure that
it is both set and true, so this suppresses an "always true" warning.