Add unique icons for each hardware shop
All the (non-racial) hardware shops used the same icon, making it hard
to distinguish between them. This change makes it possible to identify
each one visually, without needing to check the image tooltip.

I have tried to keep to the "hardware" theme as much as possible. They
use the same color scheme (cyan with black border) and have that 1990s
feel.

Modified the "J" in "NIJ" on the Nijarin Beacon so that it matches the
"J" in the Accelerated Systems. (Remove one pixel from the top serif.)