templates: more explicit conditional display
There are often large portions of a page that are displayed as a result
of a condition being met (in this case, if we've selected a ship to
compare against). This usually results in defining a whole collection
of template variables together, or not.

However, the template doesn't know about these dependencies. It only
knows that we've checked one condition (`isset($CompareShip)` in this
case), and then left a bunch of variables potentially undefined.

Until we have a more comprehensive solution for the static analysis of
templates, the best we can probably do is assert that _all_ variables
in the collection are defined. This may be a bit redundant and even
absurd in some cases, but it is certainly more explicit.

Fixes PHPStan warnings of the type:

> Parameter #1 $message of function bbifyMessage expects string, string|null given.
> Argument of an invalid type array<string, string>|null supplied for foreach, only iterables are supported.
> Parameter #1 $num of function number_format expects float, int|null given.