Remove unnecessary cache nullification
When unsetting an element in the cache, we don't need to both set the
element to null, and then unset the element. Just unsetting the element
suffices. Maybe doing both was some ancient voodoo for manipulating the
garbage collector?

Since the cache was not typehinted to accept null values, this fixes
the PHPStan warnings:

> Static property Smr\Planet::$CACHE_PLANETS (array<int, array<int, Smr\Planet>>) does not accept array<int, array<int, Smr\Planet|null>>.
> Static property Smr\Port::$CACHE_PORTS (array<int, array<int, Smr\Port>>) does not accept array<int, array<int, Smr\Port|null>>.