smr.inc.php: omit use of class_exists in saveAllAndReleaseLock
The use of `class_exists` prevents PHPStan from actually requiring that
the class exists. As a result, no static analysis is performed on the
code inside the if-block! This is unacceptable, so we forego the speed
benefit of skipping the `save*` calls in favor of making sure the calls
are correct.

This was identified because Ports were not updating at all. Since we
had not imported `\Smr\Port` properly, the `class_exists` call was
checking for the existence of a class named `\Port` instead, and thus
was always returning false.

There are comments in https://github.com/phpstan/phpstan/issues/308
suggesting that the only reason one might use `class_exists` is to
check if the class exists (as the name suggests). However, that is NOT
the only purpose of `class_exists`. It can also be used to check if the
class has been _loaded_. This is the use case for us: the class _does_
exist, we just don't want to load it here if nothing else has needed it
thus far (since its cache would be empty by definition).