history_games.php: fix use of undefined variables
When defining the "Other Info" for history games in the Summary class,
construct the database query in such a way that we can safely return
the result even if there are no players. We do this by adding IFNULL
checks on the min/max operations in the query. As a result, we can
remove the nullable from the template typehints.

Fixes PHPStan warnings of the type:

> Parameter #1 $num of function number_format expects float, int|null given.

Note that we make the query identical to the same page for non-history
games in the GameStats class. There, we remove an unnecessary ORDER BY
from the query.