HistoryPage: remove erroneous addMenu call
History pages were broken due to calling `addMenu` in both the base
class and the child classes (and thus setting `MenuItems` twice in the
template). This was leftover from the original development of the
HistoryPage class.

We need to retain the `addMenu` item call in the child classes instead
of the base class due to needing to customize the 2nd argument.