Break HistoryGames\ExtendedStats into two pages
This page had two distinct display modes: 1) listing the categories and
2) showing details for a specific category. Using the same Page class
for these two modes required having nullable template variables, which
PHPStan did not like.

Breaking this up into two pages allows us to be much more explicit
about what the expectations for the template are. Category submits are
converted to links so that we can store the category as a property of
ExtendedStatsDetail.

Fixes PHPStan warning:

> Argument of an invalid type array<string>|null supplied for foreach, only iterables are supported.