Session: fix addLink for loosely equal Pages
It was incorrect to use loose equality (==) to compare Page objects in
`Session::addLink`, because this also checks loose equality of the
property values. This resulted in unequal Pages sharing the same SN.

This bug was identified in ShopWeaponProcessor. The 3rd argument is an
`?int`; it is null when buying weapons and an int when selling weapons.
The Page for selling the first weapon (`order_id = 0`) was comparing
loosely equal to the Page for buying the same weapon (since 0 == null),
and so trying to sell your first weapon would just buy another.

The fix is to check for strict equality of the two Page objects, but
not in the sense of `===`, which also requires them to be the _same_
object. No such PHP function exists for this kind of comparison, and so
we implement it ourselves using serialization. We optimize the check by
first comparing class names, which should be unequal in most cases.
There may still be more performant alternatives to this implementation.