Location: don't store Database as a property
Storing the Database as a property of a Location object makes it hard
to compare Location objects (since the Database state is arbitrary).

It also bloats the Location object, which is an issue when storing or
dumping the object.

It could also be a problem for any persistent scripts (e.g. Discord,
NPC, etc.) if the database connection is reset, this reference would
become invalid.

The solution is to simply get the Database instance each time it is
needed.