Plotter: change how failed pathfinding is handled
* `Plotter::findDistanceToX` no longer returns false when a Path is not
  found. Instead, it will raise a `PathNotFound` exception.

* `Plotter::findReversiblePathToX` no longer throws a UserError if the
  Path is not found. Instead, it will let exceptions raise naturally.

This is done to give the callers more flexibility over how pathfinding
errors are handled, and to streamline lower-level methods in favor of
more explicit error handling in high-level methods (e.g. Page classes).

When calculating trade good transaction distances, we set the distance
to 0 if no matching transaction is found. Previously, such cases set
the distance to 1 (in a confusing way, by taking `max(1, false)`), but
this was a bit misleading. This means that exp multipliers will be 0
instead of 1 for goods with no matching trade anywhere. However, we do
need to set the minimum distance to 1 for price calculations, otherwise
goods that can't be sold anywhere could be bought for 0 credits.