Fix "Plot To Nearest" leaking Port updates
Previously, players would be able to gain knowledge about Port changes
they haven't seen yet (i.e. if the Port is not yet explored, or goods
changed since last visit) by using the "Buy/Sell Goods" category of the
"Plot To Nearest" feature, because it used the up-to-date port info.
Now, it will use the cached port info for that player, so it only uses
information that the player already has.

This fix is implemented by using the `{has,get}CachedPort` methods in
Sector::hasX instead of `{has,get}Port` when a player is supplied.