Player: revert misuse of null plottedCourse
This is an optimization fix. No functionality changes.

When we first call `Player::getPlottedCourse`, we cache the result of
the database query, even if the Player has no plotted course, so that
we don't need to perform the query again on the next call. To check if
we've set the cache, we check `isset($this->plottedCourse)`; however,
this returns false for both uninitialized _and_ null (a deficiency in
the PHP language, see https://github.com/php/php-src/pull/7029).

Therefore, we must use `false` as the "already queried but no value"
state (barring the addition of lazy-loading/caching infrastructure).

This reverts b9ad390cffa, which was causing the database query to be
repeated many times in, e.g., displaying the Local/Galaxy map.

We add a PHPStan assertion to `Player::hasPlottedCourse` to narrow the
return type of `Player::getPlottedCourse`.