Ship: add assertion to moveWeaponUp
We need to assert that `$this->weapons` is a non-empty-array, since
`array_shift` returns null if the array is empty, and we cannot append
null to `$this->weapons`. Since `moveWeaponUp` should never be called
when there are no weapons, we throw in this case.

Note that there is a bug with `array_shift` in PHPStan, so we need to
assign the result to a temporary variable for the type narrowing to
work properly. See https://github.com/phpstan/phpstan/issues/7844.

Fixes PHPStan warning:

> Property Smr\AbstractShip::$weapons (array<int, Smr\Combat\Weapon\Weapon>) does not accept array<int, Smr\Combat\Weapon\Weapon|null>.