Improve PHPStan in construction from database records
PHPStan does not track relationships between variables, so it does not
remember that we use a bool to store if `$dbRecord === null`.

Therefore, use the actual check against null as the conditional to see
if we can use `$dbRecord` as an object.

Fixes PHPStan warnings in constructors of the form:

> Cannot call method getInt() on Smr\DatabaseRecord|null.