Share DISABLE_PHPDI_COMPILATION definition
This environment variable will no longer be hard-coded to true for the
`smr-dev` service. Instead, it will take its value from the `.env` file
just like the `smr` service. Note that it is not *recommended* to use
false in the `smr-dev` service, since compiled classes will not update
as expected with local file changes until the container is rebuilt.

As an added convenience, this allows the `smr` and `smr-dev` services
to share the `environment` field.