docker-compose.yml: move flyway options to env
Specify the flyway options as environment variables instead of on the
command line. This allows us to easily change the command (e.g. `info`
instead of `migrate`) without replicating all the command-line options.

One downside is that we won't get a user-friendly error message if we
use an unsupported environment variable, whereas with command-line
arguments we would. Should be obvious enough if it breaks, though.