Fix combat results template for CombatSimulator
As a result of adding the "minimal display" option to player vs player
combat in 93c6e67e01e6b, we need to add a couple template variables to
the CombatSimulator display page.

We rework the TraderTeamCombatResults template to allow `$ThisPlayer`
to be null, in which case we don't display how much damage was done to
"you".