Add infrastructure for rector
Rector is a refactoring tool. It can help reduce the manual effort
needed for upgrading to new PHP versions or other major libraries.
We add the basic infrastructure for its use, though it will take some
time to get it to the point where we would consider using it in CI.

Note: this has an issue where editing the files change the file owner
to root, and then the files become read-only in the local filesystem.
This is because the container is running as root, and rector does not
do any special handling to help this situation (unlike, e.g., phpcbf).
Fortunately, git will revert ownership when it modifies the file.