Require use statements for all referenced names
This ensures that we declare all our dependencies up front, which is
helpful for understanding what is being used where, and enforces some
consistency.

Implemented via rector's `importNames`. Note that we explicitly only
use this for code names, and not docstring names.

I could not find a PHP_CS rule to enforce this.