Do not store Database as a class property (#1543)
Storing the Database object makes it difficult to use these classes in
persistent scripts (e.g. Discord, NPC). If the database connection is
reset, this reference would become invalid. It also makes unit testing
harder since we can't inject it as a dependency.

The solution is to simply get the Database instance each time it is
needed.