Database: add escapeNullableString method
This replaces the second parameter of the `escapeString` method. It is
more user-friendly to have a separate method for nullable input than to
overload one method with two purposes.

Furthermore, there's no way to tell PHPStan that the first argument of
`escapeString` is only allowed to be null if the 2nd argument is true.

Fixes PHPStan warning:

> Parameter #1 $string of method mysqli::real_escape_string() expects string, string|null given.