PlayerLink: check nullable attributes
Make sure that the Message author exists before trying to use its `id`.
It's not clear to me what Discord conditions result in a Message not
having an author, so it is likely pathological. Raise an Exception if
this happens.

Fixes PHPStan warning:

> Cannot access property $id on Discord\Parts\User\User|null.

The Message always has a `channel`, but the typehint in DiscordPHP was
wrong. This will be fixed on the next release of DiscordPHP, so we
ignore the warning for now.