Template: refactor middle_panel ajax logic
PHPStan cannot remember that `$doAjaxMiddle` is false if `$mid` is
null, so it doesn't understand that `$mid` is guaranteed to be a
DOMElement when used as such.

That said, it's probably more straightforward for humans as well to
scope that entire logic inside a `$mid !==` null check.

Fixes PHPStan warnings:

> Parameter #1 $node of closure expects DOMNode, DOMElement|null given.
> Cannot call method getAttribute() on DOMElement|null.