Sector: add PHPStan equality assertions
The methods `hasCachedPort` and `hasFriendlyForces` return false if a
null argument is given. We need a way to express this relationship with
PHPStan. Two annotations that did NOT work:

1. @return ($player is null ? false : bool)

This seems to express the relationship most clearly, but unfortunately
it did not actually narrow the type of `$player`.

2. @phpstan-assert-if-true !null $player

This had the surprising behavior of also asserting the inverse -- that
if the function returned false, then `$player` was not null, which we
obviously do not want. Presumably because of how "assertions" are used
traditionally (in a simple function with no complex logic), this is a
desirable side effect (https://github.com/phpstan/phpstan/issues/8351).

As alluded to in the linked PHPStan issue, we can achieve the desired
behavior by exploiting "equality assertions", for which the inverse is
not also asserted. With this, we can narrow the type of `$player` to
AbstractPlayer (and thus `!null`) if the function returns true. See
https://psalm.dev/docs/annotating_code/assertion_syntax/#equality-assertions.

Fixes from SectorMap.inc.php:

> Parameter #1 $player of method Smr\Sector::getCachedPort() expects Smr\AbstractPlayer, Smr\Player|null given.

And from SectorsFile.php:

> Parameter #1 $player of method Smr\Sector::getFriendlyForces() expects Smr\AbstractPlayer, Smr\AbstractPlayer|null given.