NpcManage: use game ID 0 for no valid games
This allows the page to display even if there are no valid games for
which NPCs can be created, since we expect the game ID to be an int,
and there is never a real game ID of 0.

Alternatively, we could have skipped a lot of the logic if the game ID
was null, but that is a more involved refactor (which creates other
problems with conditionally defined template variables).

Fixes PHPStan warnings:

> Parameter #1 $selectedGameID of class Smr\Pages\Admin\NpcManageAddAccountProcessor constructor expects int, int|null given.
> Parameter $selectedGameID of class Smr\Pages\Admin\NpcManageProcessor constructor expects int, int|null given.
> Parameter #2 $gameID of static method Smr\AbstractPlayer::getPlayer() expects int, int|null given.