Template: improve getInnerHTML closure
If we define the closure to take a `DOMElement` instead of a `DOMNode`
(the parent class), then we can be sure that it has a non-null
`ownerDocument` property.

This fixes the following PHPStan warning:

> Cannot call method saveHTML() on DOMDocument|null.

Or rather, it _should_, but the PHPStan DOM stubs are not working
properly. See https://github.com/phpstan/phpstan/issues/8837.

Note also that `childNodes` of any DOMNode must also have a non-null
`ownerDocument` (since child nodes cannot be the root node), but this
is hard to codify in a stub, because there's no way to declare a type
that is any type inheriting from `DOMNode` _except_ for `DOMDocument`.
So the previous code is valid, it just couldn't be expressed properly
to PHPStan.