DatabaseCleanup: group constructor arguments
These constructor arguments come as a package - either they are all set
or they are all null. It's tricky to demand this relationship when they
are all separately nullable arguments. This suggests that they should
be grouped (ideally in a data class, but we use an array for simplicity
here, which is made much safer with PHPStan array shapes).

Fixes PHPStan warning:

> Parameter #1 $bytes of closure expects int, int|null given.