NewsReadAdvanced: fix PHPStan warnings
The constructor arguments are only nullable under certain conditions,
due to relationships with other args that are cumbersome to enumerate.
This suggests that the class needs to be refactored, but until then,
using an empty array for some of the arguments fixes the following
PHPStan warnings:

> Parameter #1 $array of method Smr\Database::escapeArray() expects array, array<int>|null given.
> Offset 0 does not exist on array<int>|null.
> Offset 1 does not exist on array<int>|null.

This is kind of a cop out, as it doesn't address the underlying dangers
in the design, but it does placate PHPStan.