Use prepared statements for database
Instead of manually escaping variables inline in SQL statements, we now
use prepared statements. These automatically take care of escaping, and
are generally considered a best practice. It's easier to read as well,
since we have a structured array of parameters instead of a long string
concatenation. This also enables SQL static analysis using phpstan-dba.

We switch the database access layer from MySQLi to Doctrine/DBAL with
PDO as the underlying driver. This change was made after consideration
of the following desirements:

1. Named parameters (much safer and more readable than `?`).

2. Emulated prepares (relays with the database once instead of twice,
   which is important for us because communication with the database
   is our biggest runtime expense).

3. Array escaping.

4. Ability to safely close the connection (for daemon processes).

MySQLi: 4
PDO: 1, 2
Doctrine\DBAL: 1, 2 (with PDO driver), 3, 4

In Doctrine\DBAL, parameter types are passed in a separate array to aid
the driver in properly escaping the values. Unfortunately, this doesn't
validate the PHP type; instead, it does an unsafe cast of the value to
that type (e.g. if the string "3ab" is passed where an int is expected,
then the int `3` will be used). For this reason, we will continue to
validate types manually using the escaping methods. Then we can map the
PHP types to SQL types automatically using `Database::parameterTypes`.

Note that the escaping methods no longer do any actual escaping. They
serve as a simple type check in most cases (to make sure we're passing
the type of value that we expect), but in some cases they convert a PHP
type into a corresponding SQL type (e.g. escapeBoolean converts from a
PHP bool to our TRUE/FALSE enum in SQL). In the future, these methods
will be renamed, but are retained here with misleading names to reduce
the magnitude of the diff.

In data classes (e.g. Account, Player, etc.), we change the SQL string
property into a constant SQL string with named parameters, and then a
`SLQID` array with the per-instance value of those parameters. This
retains the ability to conveniently and succinctly select that instance
in SQL statements.

Some complex/conditional queries had to be reorganized so that they
could be properly analyzed by phpstan-dba.