Database: restrict escapeNumber type to int|float
Rather than manually perform an `is_numeric` assert on the input, we
now use a more strict typehint (`int|float` instead of `mixed`). As a
result, we no longer accept `numeric-string`. While not strictly
necessary, as the result is only used in a SQL query, it does help to
identify places where we have mishandled the input. (Note that we could
have used a `@phpstan-assert` instead, and retained the explicit
assertion, but this seems more straightforward.)

After making this change, PHPStan identified many places where we were
not passing in a `float|int`, and so we fix all those cases. They are
all pretty self-explanatory (e.g. check builtin functions that can
return false, replace `is_numeric` check with `filter_var` to convert
numeric-string to int, etc.).