Simplify player note database storage
Use `escapeObject` instead of `escapeBinary` so that we don't have to
do the compression manually.

Store the raw text in the database, and leave the escaping for if/when
we display it on a page.

This fixes adding double page breaks for each new line.

Existing player notes will no longer load correctly, since the database
storage has changed. Therefore, this should only be changed between
games.