NpcActor: delete plotted course on startup
If an NPC shut down uncleanly, it is possible that they still have a
plotted course set. Furthermore, if they were just killed, they might
be in the wrong sector, and will trigger an error when they try to move
to the next sector on the course.

Even absent other bugs in NPC shutdown/locking, this change helps to
ensure a clean startup due to unforeseen circumstances.