Database: add update wrapper
`Doctrine\DBAL\Connection` has an `update` method, which provides a
significantly simplified interface for a basic table update.

We pass in an array mapping table columns to PHP values, as well as
an array of WHERE equality conditions (again, mapping table columns to
PHP values). If the WHERE condition needs anything more complex than
equality, we have to construct the query ourselves.

Our `Database` class then provides a simple wrapper for this function.