Remove unnecessary parameter escaping
For Database wrapper methods delete/insert/replace/update.

The values passed in for each field map directly to columns in a table,
which we know the SQL type for exactly. The phpstan-dba extension will
be able to tell us if the passed in value is a valid PHP type for the
corresponding column. Therefore, we no longer need to explicitly state
our expectation for the PHP type using the `Database::escape*` methods.

Note that we still need to use `Database::escape*` for methods that do
more than trivial type-checking (e.g. converting a PHP boolean into the
TRUE/FALSE enum strings that we use for booleans in SQL).

Cases using `Database::escapeNullableString` have had their behavior
changed (they no longer convert an empty string to null). This is not
an issue because we manually handle the empty string in any place that
it matters (e.g. when unsetting a Discord ID or IRC Nick).