docker-compose.yml: add healthcheck to mysql service
This makes sure that the mysql server is accepting connections before
we attempt to use it. Otherwise the service that depends on it will
fail if it doesn't have its own retry mechanism (e.g. flyway does, but
phpstan/phpunit do not).

Note that we don't use the common "mysqladmin ping" solution, since
that returns success before the mysql server is actually ready.