Replace GROUP BY hack with null result check
Don't (ab)use the `GROUP BY` SQL clause on a single-valued column in
the result set as a hack to avoid null result values from aggregate
functions (e.g. `MAX`).

Then, to determine if no rows matched the query conditions, we simply
check if the result value is null instead of checking if the result
set is empty.

This replaces the GROUP BY logic added in d25f139870a21.