Account: fix HOF user score SQL
We cannot pass `hof_types` as a string (from an imploded array) to a
prepared statement, because the entire string will be escaped as part
of the parameter substitution. Therefore, we need to pass it in array
form, which means `Account::getUserScoreCaseStatement` needs to return
the raw array instead of the imploded array.

The imploded type does need to be manually escaped when used in the
case statement, since `Database::escapeString` no longer actually
does any escaping (it's just a type-check now).