Make HOF query resolvable in pure SQL
In order to analyze the query with phpstan-dba, we need the `game_id`
conditional to be in SQL instead of PHP (as long as we are executing
the query in a single `Database::read` call).

Therefore, instead of checking nullability of `game_id` in PHP, we use
the MySQL `IF` function.