FeatureRequest: make query resolvable
In order for the query to be resolvable, we need to use conditional
logic inside the query itself, rather than in PHP. We do this by only
checking the `posting_time` if the query is not in the "New" category.
Without this change, we cannot analyze the SQL query in phpstan-dba.

Add a `CATEGORY_NEW` class constant to avoid repeating the literal
string value "New".