channel_msg_money.php: use IFNULL with SUM in queries
Most aggregate functions (including SUM) always return a single record,
so there is no need to check if the record exists. Furthermore, if the
return set has no rows, SUM returns null, so if we don't want the value
to be nullable, we need an IFNULL check.

Fixes phpstan-dba warnings of the type:

> Smr\DatabaseRecord::getInt returns type int, but field `total_onship` has type int|null