Fix issues with strict equality comparison
PHPStan has identified places where switching to strict equality
comparison has resulted in always-true or always-false scenarios. For
example, if we strictly compare a float to "0", it will always return
false because "0" is an int (whereas "0.0" is the appropriate float).

* In `get_colored_text_range`, require that the min/max values are
  ints instead of floats. We don't need this flexibility and it makes
  the equality comparison more robust.

* When comparing a numeric-string to an int in `smrBBCode` "race"
  condition, cast to an int before comparing to an int.

More issues with strict comparison may yet arise, but they will not be
discovered by static analysis.