Port: add method isCreditedAttacker
We have to be careful when checking if an object is in an array, since
strict equality comparison of objects looks at reference equality.
Therefore, to more safely check if a player is in an array of port
attacker players, we add the method Port::isCreditedAttacker to use the
Player::equals method under the hood for equality comparison.