Player: add isObserver method
We were often repeating some non-trivial logic to check if a Player was
in the Globals::getHiddenPlayers array. To avoid duplication, we add a
Player::isObserver method to perform this check more easily.

Note that the changes to the Player "getter" methods will always make
the Player object now, whereas it was skipped if the player was hidden.
In theory this is potentially less efficient, but in practice we never
have any hidden players.