Improve combat typehints
Add a bunch of (very complex) type aliases in the PHPStan configuration
so that we can reuse these combat-related types in several places.

Previously, we were using "mixed" to avoid duplicating these massive
type strings. We want to avoid "mixed" whenever possible, because it
inhibits static type analysis.

Some additional runtime assertions are needed since we are reusing
typehints for both regular weapons and combat drones, which are arrays
with different keys. Refactoring or switching to class objects instead
of arrays would probably improve this situation.