First step in generalizing combat interface
In the combat result arrays, we were using keys like "TargetPort" or
"TargetPlayer" etc. to store the target. This makes it impossible to
reuse the type alises between target types. We change the key to just
"Target" to make consolidation in the future easier.

Note that to take advantage of this change, we will need to add a base
class for combatants, so that we don't need to specify Port/Player/etc.
for the type of the "Target" field.