Use only booleans in if conditions
Fixes phpstan-strict-rules warning, e.g.:

> Only booleans are allowed in an if condition, string given.

It is better to be explicit about what condition we're actually trying
to test, which then allows PHPStan to more easily check the effects of
that condition (and if it's even possible). It should also make reading
the code easier, since it provides a clearer indication of what type a
variable can be.