Disallow use of the empty function
SlevomatCodingStandard.ControlStructures.DisallowEmpty

Because `empty` also checks if a value exists, static analysis cannot
warn when the variable you're checking is expected to exist. This makes
code using `empty` much more vulnerable to typos.

In some cases, `empty` seems to be a more efficient way to express the
logic that we actually care about, but I'm willing to trust that the
benefits of these changes will outweigh the convenience of `empty`.